chore(checkbox-input): move to packages
