+++ packages/components/inputs/checkbox-input/src/checkbox-input.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { screen, render } from '../../../../test/test-utils';
+import { screen, render } from '../../../../../test/test-utils';
 import CheckboxInput from './checkbox-input';
 
 jest.mock('tiny-invariant');
