+++ packages/components/inputs/checkbox-input/src/checkbox-input.story.js
@@ -4,8 +4,8 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text } from '@storybook/addon-knobs/react';
 import { Value } from 'react-value';
 import Spacings from '@commercetools-uikit/spacings';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import CheckboxInput from './checkbox-input';
 
 storiesOf('Components|Inputs', module)
