+++ packages/components/inputs/checkbox-input/src/checkbox-input.styles.js
