+++ packages/components/inputs/checkbox-input/src/checkbox-input.visualspec.js
