+++ packages/components/inputs/checkbox-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './checkbox-input';
+
+export { default as version } from './version';
