--- src/components/inputs/checkbox-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './checkbox-input';
